feat(state): extract awards operations
