+++ src/main/scala/scatan/model/game/ops/ScoreOps.scala
@@ -4,6 +4,7 @@ import scatan.model.components.*
 import scatan.model.components.AssignedBuildingsAdapter.asPlayerMap
 import scatan.model.game.ScatanState
 import scatan.model.game.config.ScatanPlayer
+import scatan.model.game.ops.AwardOps.*
 
 object ScoreOps:
   extension (state: ScatanState)
